save point added to prefab and ready for been used

+++ Assets/Scripts/SavingSystem/SaveLoad.cs
@@ -64,7 +64,8 @@ public class SaveLoad {
 	
 	public static  void FirstGame () {
 		//cancel prevoius saved game and saving points
-		File.Delete(SAVING_PATH);
+		if(File.Exists(SAVING_PATH))File.Delete(SAVING_PATH);
+		if(File.Exists(SavingPoints.SAVING_POINTS_PATH))File.Delete(SavingPoints.SAVING_POINTS_PATH);
 		Debug.Log ("New Game");
 		savedGame.x = -22.30036f;
 		savedGame.y = -2.331829f;
